[BPM-225] 질문 댓글 및 대댓글 로직 수정

+++ presentation/src/main/java/com/team/bpm/presentation/ui/main/community/question_detail/QuestionDetailViewModel.kt
@@ -271,14 +271,14 @@ class QuestionDetailViewModel @Inject constructor(
 
     private fun onClickCommentActionButton(
         selectedCommentId: Int,
-        selectedCommentAuthorId: Int,
+        selectedCommentAuthorId: Long,
         parentCommentId: Int?
     ) {
         viewModelScope.launch {
             _state.update {
                 val bottomSheetButtonList = mutableListOf<BottomSheetButton>().apply {
                     add(BottomSheetButton.REPLY_COMMENT)
-                    if (selectedCommentAuthorId.toLong() == state.value.userId) {
+                    if (selectedCommentAuthorId == state.value.userId) {
                         add(BottomSheetButton.DELETE_COMMENT)
                     } else {
                         add(BottomSheetButton.REPORT_COMMENT)
